+++ src/APIs/GauntletManagerAPI.hpp
@@ -48,9 +48,7 @@ public:
     web::WebFuture remove(int id);
     web::WebFuture setFeatured(int id);
     web::WebFuture uploadIcon(std::filesystem::path const& path);
-    web::WebFuture fetchManagers();
-    web::WebFuture fetchHelpers();
-    web::WebFuture fetchSupporters();
+    web::WebFuture fetchRoles();
 
     bool isLoggedIn() const;
     void setToken(std::string const& token);