+++ src/APIs/GauntletManagerCache.hpp
@@ -2,6 +2,7 @@
 #include <Geode/Geode.hpp>
 #include <Geode/utils/web.hpp>
 #include <Geode/utils/async.hpp>
+#include <chrono>
 
 using namespace geode::prelude;
 
@@ -20,20 +21,23 @@ public:
 private:
     GauntletManagerCache();
     struct RoleCache {
-        std::function<web::WebFuture()> fetch;
         std::string jsonKey;
         std::string label;
 
         std::unordered_set<int> ids;
         bool hasFetched = false;
         std::vector<std::pair<int, std::function<void(bool)>>> waiting;
-        async::TaskHolder<web::WebResponse> request;
     };
 
-    void fetchRole(RoleCache& cache);
+    void fetchRoles();
     void checkRole(RoleCache& cache, int accountID, std::function<void(bool)> callback);
 
     RoleCache m_managers;
     RoleCache m_helpers;
     RoleCache m_supporters;
+
+    bool m_isFetching = false;
+    async::TaskHolder<web::WebResponse> m_request;
+
+    std::chrono::steady_clock::time_point m_lastRefresh = std::chrono::steady_clock::time_point::min();
 };